+++ ktrain/text/eda.py
@@ -177,6 +177,15 @@ class TopicModel():
         return self.get_topics()
 
 
+    def get_document_topic_distribution(self):
+        """
+        Gets the document-topic distribution.
+        Each row is a document and each column is a topic
+        """
+        if not self.doc_topics: raise Exception('Must call build first.')
+        return self.doc_topics
+
+
     def get_word_weights(self, topic_id, n_words=100):
         """
         Returns a list tuples of the form: (word, weight) for given topic_id.
@@ -221,7 +230,6 @@ class TopicModel():
         show_counts(bool): If True, print topics with document counts, where
                            the count is the number of documents with that topic as primary.
         """
-
         topics = self.get_topics(n_words=n_words, as_string=True)
         if show_counts:
             self._check_build()
@@ -286,8 +294,11 @@ class TopicModel():
                         of texts supplied to self.build (which is the order of self.doc_topics).
 
         Returns:
-            list of tuples:  list of tuples where each tuple is of the form 
-                             (text, doc_id, topic_probability, topic_id).
+            list of dicts:  list of dicts with keys:
+                            'text': text of document
+                            'doc_id': ID of document
+                            'topic_proba': topic probability (or score)
+                            'topic_id': ID of topic
             
         """
         self._check_build()
@@ -296,11 +307,11 @@ class TopicModel():
         result_texts = []
         for topic_id in topic_ids:
             if topic_id not in self.topic_dict: continue
-            texts = [tup + (topic_id,) for tup in self.topic_dict[topic_id] 
-                                           if not doc_ids or tup[1] in doc_ids]
+            texts = [{'text':tup[0], 'doc_id':tup[1], 'topic_proba':tup[2], 'topic_id':topic_id} for tup in self.topic_dict[topic_id] 
+                                                                                                     if not doc_ids or tup[1] in doc_ids]
             result_texts.extend(texts)
         if not rank:
-            result_texts = sorted(result_texts, key=lambda x:x[1])
+            result_texts = sorted(result_texts, key=lambda x:x['doc_id'])
         return result_texts
 
 
@@ -319,7 +330,7 @@ class TopicModel():
                         
         """
         docs = self.get_docs(topic_ids=topic_ids, doc_ids=doc_ids)
-        return np.array([self.doc_topics[idx] for idx in [x[1] for x in docs]])
+        return np.array([self.doc_topics[idx] for idx in [x['doc_id'] for x in docs]])
 
 
     def get_texts(self,  topic_ids=[]):
@@ -635,7 +646,7 @@ class TopicModel():
         results = []
         for i in mb:
             for doc in progress_bar(docs, parent=mb):
-                text = doc[0]
+                text = doc['text']
                 if not case_sensitive: text = text.lower()
                 matches = pattern.findall(text)
                 if matches: results.append(doc)